# backend/src/main/resources/openapi/registry-api.yaml
openapi: 3.0.3
info:
  title: HeatMap Registry API
  description: Реестр AI-инструментов PDLC
  version: 0.1.0
servers:
  - url: /
...
security: []
paths:
...
  /api/tools:
    get:
      tags: [Tools]
      operationId: listTools
      parameters:
        - name: tab
          in: query
          schema: { type: string, default: TOP }
        - name: role
          in: query
          schema: { type: array, items: { type: string } }
        - name: framework
          in: query
          schema: { type: array, items: { type: string } }
        - name: constraints
          in: query
          schema: { type: array, items: { type: string } }
        - name: toolType
          in: query
          schema: { type: array, items: { type: string } }
        - name: search
          in: query
          schema: { type: string }
        - name: sort
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Список инструментов
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ToolResponse' }
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [Tools]
      operationId: createTool
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateToolRequest'
      responses:
        '201':
          description: Инструмент создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolResponse'
        default:
          $ref: '#/components/responses/Error'
  /api/tools/counts:
    get:
      tags: [Tools]
      operationId: getToolCounts
      responses:
        '200':
          description: Счётчики по вкладкам реестра
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolCountsResponse'
        default:
          $ref: '#/components/responses/Error'
  /api/tools/stats:
    get:
      tags: [Tools]
      operationId: getToolStats
      responses:
        '200':
          description: Общая статистика реестра
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatsResponse'
        default:
          $ref: '#/components/responses/Error'
  /api/tools/filter-options:
    get:
      tags: [Tools]
      operationId: getFilterOptions
      responses:
        '200':
          description: Варианты значений для фильтров
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FilterOptionsResponse'
        default:
          $ref: '#/components/responses/Error'
  /api/tools/mine:
    get:
      tags: [Tools]
      operationId: listMyTools
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Собственные инструменты пользователя
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ToolResponse' }
        default:
          $ref: '#/components/responses/Error'
  /api/tools/downloaded:
    get:
      tags: [Tools]
      operationId: listDownloadedTools
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Скачанные пользователем инструменты
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ToolResponse' }
        default:
          $ref: '#/components/responses/Error'
  /api/tools/{id}:
    parameters:
      - $ref: '#/components/parameters/ToolId'
    get:
      tags: [Tools]
      operationId: getTool
      responses:
        '200':
          description: Инструмент
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolResponse'
        default:
          $ref: '#/components/responses/Error'
    patch:
      tags: [Tools]
      operationId: updateTool
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateToolRequest'
      responses:
        '200':
          description: Инструмент обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolResponse'
        default:
          $ref: '#/components/responses/Error'
    delete:
      tags: [Tools]
      operationId: deleteTool
      security: [{ bearerAuth: [] }]
      responses:
        '204':
          description: Инструмент удалён
        default:
          $ref: '#/components/responses/Error'
  /api/tools/{id}/view:
    parameters:
      - $ref: '#/components/parameters/ToolId'
    post:
      tags: [Tools]
      operationId: registerToolView
      responses:
        '200':
          description: Просмотр учтён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolResponse'
        default:
          $ref: '#/components/responses/Error'
  /api/tools/{id}/download:
    parameters:
      - $ref: '#/components/parameters/ToolId'
    post:
      tags: [Tools]
      operationId: registerToolDownload
      responses:
        '200':
          description: Скачивание учтено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolResponse'
        default:
          $ref: '#/components/responses/Error'
  /api/tools/{id}/rating:
    parameters:
      - $ref: '#/components/parameters/ToolId'
    post:
      tags: [Tools]
      operationId: rateTool
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RateToolRequest'
      responses:
        '200':
          description: Оценка сохранена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolResponse'
        default:
          $ref: '#/components/responses/Error'
  /api/tools/{id}/withdraw:
    parameters:
      - $ref: '#/components/parameters/ToolId'
    post:
      tags: [Tools]
      operationId: withdrawTool
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Заявка отозвана в черновик
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolResponse'
        default:
          $ref: '#/components/responses/Error'

  # ===================== Admin: moderation & archive =====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ToolId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Error:
      description: Ошибка
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required: [status, error, message]
      properties:
        status: { type: integer }
        error: { type: string }
        message: { type: string }
        fieldErrors:
          type: object
          additionalProperties: { type: string }
        conflictToolId: { type: string, nullable: true }
        conflictToolName: { type: string, nullable: true }

    # ----- Auth -----
...
    ToolResponse:
      type: object
      required:
        - id
        - name
        - description
        - stage
        - status
        - roles
        - framework
        - constraints
        - ownerName
        - downloads
        - efficiencyPct
        - isTop
        - views
        - avgRating
        - ratingsCount
        - canManage
        - notesCount
        - createdAt
        - updatedAt
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string }
        shortDescription: { type: string, nullable: true }
        stage: { type: string }
        status: { type: string }
        roles: { type: array, items: { type: string } }
        framework: { type: array, items: { type: string } }
        toolType: { type: string, nullable: true }
        constraints: { type: array, items: { type: string } }
        sourceLabel: { type: string, nullable: true }
        ownerName: { type: string }
        downloads: { type: integer, format: int32 }
        dau: { type: integer, format: int32, nullable: true }
        efficiencyPct: { type: integer, format: int32 }
        isTop: { type: boolean }
        views: { type: integer, format: int64 }
        avgRating: { type: number, format: double }
        ratingsCount: { type: integer, format: int32 }
        myRating: { type: integer, format: int32, nullable: true }
        canManage: { type: boolean }
        rejectionReason: { type: string, nullable: true }
        notesCount: { type: integer, format: int64 }
        segment: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    RateToolRequest:
      type: object
      required: [rating]
      properties:
        rating:
          type: integer
          format: int32
          x-field-extra-annotation: '@field:NotNull(message = "Укажите оценку") @field:Min(1, message = "Оценка - от 1 до 5") @field:Max(5, message = "Оценка - от 1 до 5")'
    CreateToolRequest:
      type: object
      required: [name, description, roles]
      properties:
        name:
          type: string
          maxLength: 255
          x-field-extra-annotation: '@field:jakarta.validation.constraints.NotBlank(message = "Введите название инструмента")'
        shortDescription:
          type: string
          nullable: true
          x-field-extra-annotation: '@field:Size(max = 300, message = "Не более 300 символов")'
        description:
          type: string
          x-field-extra-annotation: '@field:jakarta.validation.constraints.NotBlank(message = "Введите описание")'
        roles:
          type: array
          items: { type: string }
          x-field-extra-annotation: '@field:jakarta.validation.constraints.NotEmpty(message = "Выберите хотя бы одну роль")'
        framework: { type: array, items: { type: string }, default: [] }
        toolType:
          type: string
          nullable: true
          x-field-extra-annotation: '@field:Size(max = 60, message = "Не более 60 символов")'
        constraints: { type: array, items: { type: string }, default: [] }
        sourceLabel:
          type: string
          nullable: true
          x-field-extra-annotation: '@field:Pattern(regexp = "^(?i)https?://(?:[\\w-]+\\.)*(sc-ci|sbrf-bitbucket|stash|confluence|jira|mapp|sbertrack|onework)[\\w.-]*(?::\\d+)?(/.*)?$", message = "Ссылка должна вести на корпоративный сервис (sc-ci, sbrf-bitbucket, stash, confluence, jira, mapp, sbertrack, onework)")'
        stage: { type: string, nullable: true }
        status: { type: string, nullable: true }
    UpdateToolRequest:
      type: object
      properties:
        name: { type: string, maxLength: 255 }
        description: { type: string }
        shortDescription:
          type: string
          nullable: true
          x-field-extra-annotation: '@field:Size(max = 300, message = "Не более 300 символов")'
        roles: { type: array, items: { type: string } }
        framework: { type: array, items: { type: string } }
        toolType:
          type: string
          nullable: true
          x-field-extra-annotation: '@field:Size(max = 60, message = "Не более 60 символов")'
        constraints: { type: array, items: { type: string } }
        sourceLabel:
          type: string
          nullable: true
          x-field-extra-annotation: '@field:Pattern(regexp = "^(?i)https?://(?:[\\w-]+\\.)*(sc-ci|sbrf-bitbucket|stash|confluence|jira|mapp|sbertrack|onework)[\\w.-]*(?::\\d+)?(/.*)?$", message = "Ссылка должна вести на корпоративный сервис (sc-ci, sbrf-bitbucket, stash, confluence, jira, mapp, sbertrack, onework)")'
        stage: { type: string, nullable: true }
        status: { type: string, nullable: true }
        downloads: { type: integer, format: int32, minimum: 0 }
        dau: { type: integer, format: int32, nullable: true }
        efficiencyPct: { type: integer, format: int32, minimum: 0, maximum: 100 }
        segment:
          type: string
          nullable: true
          x-field-extra-annotation: '@field:Size(max = 128, message = "Не более 128 символов")'
    ToolCountsResponse:
      type: object
      required: [top, access, usage, habit, standard, total]
      properties:
        top: { type: integer, format: int64 }
        access: { type: integer, format: int64 }
        usage: { type: integer, format: int64 }
        habit: { type: integer, format: int64 }
        standard: { type: integer, format: int64 }
        total: { type: integer, format: int64 }
    FilterOptionsResponse:
      type: object
      required: [roles, frameworks, constraints, toolTypes]
      properties:
        roles: { type: array, items: { type: string } }
        frameworks: { type: array, items: { type: string } }
        constraints: { type: array, items: { type: string } }
        toolTypes: { type: array, items: { type: string } }
    StatsResponse:
      type: object
      required: [totalTools, newThisWeek, accessCount, usageCount, standardCount]
      properties:
        totalTools: { type: integer, format: int64 }
        newThisWeek: { type: integer, format: int64 }
        accessCount: { type: integer, format: int64 }
        usageCount: { type: integer, format: int64 }
        standardCount: { type: integer, format: int64 }

    # ----- Admin: users -----
...